deck, app-graph, app-toolkit: fix the lint the move left behind
Unused imports the rewrites orphaned, bare "." imports the barrel move produced,
and one import order.
