+++ packages/plugins/plugin-deck/src/url/navigate.ts
@@ -16,8 +16,8 @@ import * as AttentionCapabilities from '@dxos/plugin-attention/AttentionCapabili
 
 import { CompanionViewState, DeckCapabilities } from '#types';
 
-import { Navigation } from '.';
 import { getRenderedPlanks, isCompanionOpen, resolveCompanionAnchor } from '../util';
+import * as Navigation from './navigation';
 import { projectUrl } from './project';
 
 /** What the URL currently says is open. */
